fix: move MANAGER_URLS constants from core to navbar
