+++ packages/manager-core/src/manager-core.constants.js
@@ -1,14 +1,3 @@
-export const MANAGER_URLS = {
-  default: 'https://www.ovh.com/manager/#/',
-  sunrise: 'https://www.ovh.com/manager/sunrise/index.html#/',
-  gamma: 'https://www.ovh.com/manager/sunrise/index.html#/',
-  v3: 'https://www.ovh.com/managerv3/home.pl',
-  portal: 'https://www.ovh.com/manager/portal/index.html#/',
-  partners: 'https://www.ovh.com/manager/partners/',
-  labs: 'https://www.ovh.com/manager/sunrise/uxlabs/#!/',
-};
-
-
 export const LANGUAGES = {
   available: [{
     name: 'Deutsch',
@@ -505,5 +494,5 @@ export const URLS = {
 };
 
 export default {
-  MANAGER_URLS, LANGUAGES, REDIRECT_URLS, URLS,
+  LANGUAGES, REDIRECT_URLS, URLS,
 };
