+++ packages/manager-layout-ovh/src/navbar/navbar.constants.js
@@ -1,0 +1,14 @@
+export const MANAGER_URLS = {
+  default: 'https://www.ovh.com/manager/#/',
+  sunrise: 'https://www.ovh.com/manager/sunrise/index.html#/',
+  gamma: 'https://www.ovh.com/manager/sunrise/index.html#/',
+  v3: 'https://www.ovh.com/managerv3/home.pl',
+  portal: 'https://www.ovh.com/manager/portal/index.html#/',
+  partners: 'https://www.ovh.com/manager/partners/',
+  labs: 'https://www.ovh.com/manager/sunrise/uxlabs/#!/',
+};
+
+
+export default {
+  MANAGER_URLS,
+};
